+++ x-pack/solutions/observability/plugins/slo/public/components/manage_linked_dashboards/manage_linked_dashboards_flyout.tsx
@@ -5,9 +5,7 @@
  * 2.0.
  */
 
-import type { Criteria, EuiBasicTableColumn } from '@elastic/eui';
 import {
-  EuiBasicTable,
   EuiButton,
   EuiButtonEmpty,
   EuiFlexGroup,
@@ -16,20 +14,15 @@ import {
   EuiFlyoutBody,
   EuiFlyoutFooter,
   EuiFlyoutHeader,
-  EuiLink,
-  EuiSearchBar,
-  EuiText,
   EuiTitle,
   useGeneratedHtmlId,
 } from '@elastic/eui';
-import { DASHBOARD_APP_LOCATOR } from '@kbn/deeplinks-analytics';
 import { i18n } from '@kbn/i18n';
 import type { Asset } from '@kbn/slo-schema';
-import { debounce } from 'lodash';
 import React, { useState } from 'react';
-import type { DashboardLocatorParams } from '@kbn/dashboard-plugin/common';
-import { useKibana } from '../../hooks/use_kibana';
-import { useFetchDashboards } from './hooks/use_fetch_dashboards';
+import { DashboardSearchTable } from './dashboard_search_table';
+import type { Dashboard } from './types';
+import { AssignedDashboardList } from './assigned_dashboard_list';
 
 interface Props {
   assets: Asset[];
@@ -37,19 +30,9 @@ interface Props {
   onSave: () => void;
 }
 
-interface Dashboard {
-  id: string;
-  title: string;
-}
-
 export function ManageLinkedDashboardsFlyout({ assets, onClose, onSave }: Props) {
-  const {
-    services: { share },
-  } = useKibana();
-  const [search, setSearch] = useState<string>();
-  const [page, setPage] = useState(0);
-  const debounceSearch = debounce((value: string) => setSearch(value), 300);
-  const [selectedDashboards, setSelectedDashboards] = useState<Dashboard[]>(
+  const flyoutId = useGeneratedHtmlId({ prefix: 'linkedDashboardFlyout' });
+  const [assignedDashboards, setSelectedDashboards] = useState<Dashboard[]>(
     assets
       .filter((asset) => asset.type === 'dashboard')
       .map((asset) => ({ id: asset.id, title: asset.label }))
@@ -65,70 +48,10 @@ export function ManageLinkedDashboardsFlyout({ assets, onClose, onSave }: Props)
     );
   };
 
-  const { data, isLoading } = useFetchDashboards({ search, page: page + 1 });
-
-  const flyoutId = useGeneratedHtmlId({ prefix: 'linkedDashboardFlyout' });
-  const dashboardLocator = share.url.locators.get<DashboardLocatorParams>(DASHBOARD_APP_LOCATOR);
-
-  const columns: Array<EuiBasicTableColumn<Dashboard>> = [
-    {
-      field: 'label',
-      name: 'Dashboard',
-      render: (_, { id, title }) => (
-        <EuiLink
-          data-test-subj="sloColumnsLink"
-          href={dashboardLocator?.getRedirectUrl({ dashboardId: id } ?? '')}
-          target="_blank"
-        >
-          {title}
-        </EuiLink>
-      ),
-    },
-    {
-      field: 'action',
-      name: 'Action',
-      render: (_, dashboard: Dashboard) =>
-        selectedDashboards.find((asset) => asset.id === dashboard.id) ? (
-          <EuiButton
-            data-test-subj="unassignButton"
-            onClick={() => unassign(dashboard)}
-            size="s"
-            color="danger"
-          >
-            {i18n.translate('xpack.slo.columns.unassignButtonLabel', {
-              defaultMessage: 'Unassign',
-            })}
-          </EuiButton>
-        ) : (
-          <EuiButton
-            data-test-subj="assignButton"
-            onClick={() => assign(dashboard)}
-            size="s"
-            color="success"
-          >
-            {i18n.translate('xpack.slo.columns.assignButtonLabel', { defaultMessage: 'Assign' })}
-          </EuiButton>
-        ),
-    },
-  ];
-
-  const onTableChange = ({ page: newPage }: Criteria<Dashboard>) => {
-    if (newPage) {
-      const { index } = newPage;
-      setPage(index);
-    }
-  };
-
-  const pagination = {
-    pageIndex: page,
-    pageSize: 25,
-    totalItemCount: data?.total ?? 0,
-  };
-
   return (
-    <EuiFlyout onClose={onClose} aria-labelledby={flyoutId} ownFocus>
+    <EuiFlyout onClose={onClose} aria-labelledby={flyoutId} ownFocus size="s">
       <EuiFlyoutHeader hasBorder>
-        <EuiTitle>
+        <EuiTitle size="s">
           <h2 id={flyoutId}>
             {i18n.translate(
               'xpack.slo.manageLinkedDashboardsFlyout.managedLinkedDashboardsTitleLabel',
@@ -140,32 +63,14 @@ export function ManageLinkedDashboardsFlyout({ assets, onClose, onSave }: Props)
         </EuiTitle>
       </EuiFlyoutHeader>
       <EuiFlyoutBody>
-        <EuiFlexGroup direction="column">
-          <EuiText size="s">
-            {i18n.translate('xpack.slo.manageLinkedDashboardsFlyout.description', {
-              defaultMessage: 'Select dashboards which you want to add and assign to this SLO.',
-            })}
-          </EuiText>
-          <EuiFlexItem grow>
-            <EuiSearchBar
-              query={search}
-              onChange={({ queryText }) => {
-                debounceSearch(queryText);
-              }}
-            />
-          </EuiFlexItem>
-
-          <EuiBasicTable
-            compressed
-            columns={columns}
-            itemId="id"
-            items={data?.results ?? []}
-            loading={isLoading}
-            pagination={pagination}
-            onChange={onTableChange}
+        <EuiFlexGroup direction="column" gutterSize="m">
+          <DashboardSearchTable
+            assignedDashboards={assignedDashboards}
+            assign={assign}
+            unassign={unassign}
           />
 
-          <pre>{JSON.stringify(selectedDashboards, null, 2)}</pre>
+          <AssignedDashboardList dashboards={assignedDashboards} unassign={unassign} />
         </EuiFlexGroup>
       </EuiFlyoutBody>
       <EuiFlyoutFooter>
