+++ x-pack/solutions/observability/plugins/slo/public/components/manage_linked_dashboards/types.ts
@@ -4,3 +4,8 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+
+export interface Dashboard {
+  id: string;
+  title: string;
+}
